refactor: create a pipeline lib
