+++ tasks/show_linters_issues/show_linters_issues.py
@@ -5,7 +5,7 @@ Invoke tasks to fix the linter
 from invoke import task
 from invoke.exceptions import Exit
 
-from tasks.libs.pipeline_notifications import GITHUB_SLACK_MAP
+from tasks.libs.pipeline.notifications import GITHUB_SLACK_MAP
 from tasks.show_linters_issues.golangci_lint_parser import (
     count_lints_per_team,
     display_nb_lints_per_team,
